refactor: rename extension CORS plugin module
