+++ packages/vite-plugin/src/node/index.ts
@@ -2,7 +2,7 @@ import type { PluginOption } from 'vite'
 import type { ManifestV3Export } from './defineManifest'
 import { pluginOptionsProvider } from './plugin-optionsProvider'
 import { pluginBackground } from './plugin-background'
-import { pluginExtensionCors } from './extensionCors'
+import { pluginExtensionCors } from './plugin-extensionCors'
 import { pluginContentScripts } from './plugin-contentScripts'
 import { pluginContentScriptsCss } from './plugin-contentScripts_css'
 import { pluginDeclaredContentScripts } from './plugin-contentScripts_declared'
