+++ packages/vite-plugin/src/node/plugin-extensionCors.ts
